feat: implement full_procedure script, logging setup and integrate into NSA scripts

+++ scripts/nsa/drop_dates_with_duplicated_documents.py
@@ -1,6 +1,12 @@
 import pandas as pd
 import pymongo
+from tqdm import tqdm
 import typer
+
+from loguru import logger
+from juddges.utils.logging import setup_loguru
+
+setup_loguru(extra={"script": __file__})
 
 
 def main(
@@ -16,7 +22,17 @@ def main(
     duplicated = df[df.duplicated(subset="document_ids", keep=False)]
     dates_to_drop = duplicated["date"].unique()
 
-    for date in dates_to_drop:
+    logger.info(f"Found {len(dates_to_drop)} dates with duplicated documents.")
+
+    if len(dates_to_drop) == 0:
+        logger.info("No dates with duplicated documents found.")
+        return
+
+    if not typer.confirm("Do you want to delete them?"):
+        logger.info("Aborted.")
+        raise typer.Abort()
+
+    for date in tqdm(dates_to_drop):
         dates_col.delete_many({"date": date})
